refactor: isolate session toolbar filter metadata
